+++ R/misclassification.R
@@ -1,6 +1,8 @@
-#' Sensitivity analysis for misclassification.
+#' Sensitivity analysis for disease or exposure misclassification.
 #'
-#' Simple sensitivity analysis for misclassification.
+#' Simple sensitivity analysis for disease or exposure misclassification.
+#' Confidence interval for odds ratio is computed as in Chu et al. (2006),
+#' for exposure misclassification.
 #'
 #' @param case Outcome variable. If a variable, this variable is tabulated against.
 #' @param exposed Exposure variable.
@@ -9,30 +11,33 @@
 #' \item exposure: bias analysis for exposure misclassification; corrections using
 #' sensitivity and specificity: nondifferential and independent errors,
 #' \item outcome: bias analysis for outcome misclassification.
-#' \item confounder: bias analysis for confounder misclassification.
 #' }
-#' @param bias Deprecated, please use bias_parms instead.
 #' @param bias_parms Vector defining the bias parameters. This vector has 4 elements
 #' between 0 and 1, in the following order:
 #' \enumerate{
-#' \item Sensitivity of exposure (or outcome, or confounder) classification among those with the outcome,
-#' \item Sensitivity of exposure (or outcome, or confounder) classification among those without the outcome,
-#' \item Specificity of exposure (or outcome, or confounder) classification among those with the outcome,and
-#' \item Specificity of exposure (or outcome, or confounder) classification among those without the outcome.
+#' \item Sensitivity of exposure (or outcome) classification among those with the outcome (or exposure),
+#' \item Sensitivity of exposure (or outcome) classification among those without the outcome (or exposure),
+#' \item Specificity of exposure (or outcome) classification among those with the outcome (or exposure), and
+#' \item Specificity of exposure (or outcome) classification among those without the outcome (or exposure).
 #' }
 #' @param alpha Significance level.
 #' 
 #' @return A list with elements:
-#' \item{obs.data}{The analysed 2 x 2 table from the observed data.}
+#' \item{obs.data}{The analyzed 2 x 2 table from the observed data.}
 #' \item{corr.data}{The expected observed data given the true data assuming
-#' misclassfication.}
+#' misclassification.}
 #' \item{obs.measures}{A table of observed relative risk and odds ratio with
 #' confidence intervals.}
-#' \item{adj.measures}{A table of adjusted relative risk and odds ratio.}
+#' \item{adj.measures}{A table of adjusted relative risk and odds ratio with confidence
+#' interval for odds ratio.}
 #' \item{bias.parms}{Input bias parameters.}
 #'
 #' @references Lash, T.L., Fox, M.P, Fink, A.K., 2009 \emph{Applying Quantitative
 #' Bias Analysis to Epidemiologic Data}, pp.79--108, Springer.
+#'
+#' Chu, H., Zhaojie, W., Cole, S.R., Greenland, S., \emph{Sensitivity analysis of
+#' misclassification: A graphical and a Bayesian approach},
+#' Annals of Epidemiology 2006;16:834-841.
 #' 
 #' @examples
 #' # The data for this example come from:
@@ -51,19 +56,21 @@
 #' nrow = 2, byrow = TRUE),
 #' type = "outcome",
 #' bias_parms = c(.53, .53, .99, .99))
+#' # The following example comes from Chu et al. Sensitivity analysis of
+#' # misclassification: A graphical and a Bayesian approach.
+#' # Annals of Epidemiology 2006;16:834-841.
+#' misclassification(matrix(c(126, 92, 71, 224),
+#' dimnames = list(c("Case", "Control"), c("Smoker +", "Smoker -")),
+#' nrow = 2, byrow = TRUE),
+#' type = "exposure",
+#' bias_parms = c(.94, .94, .97, .97))
 #' @export
 #' @importFrom stats qnorm
 misclassification <- function(case,
                               exposed,
-                              type = c("exposure", "outcome", "confounder"),
-                              bias = NULL,
+                              type = c("exposure", "outcome"),
                               bias_parms = NULL,
                               alpha = 0.05){
-    if (!missing(bias)) {
-        warning("Argument bias is deprecated; please use bias_parms instead.", 
-                call. = FALSE)
-        bias_parms <- bias
-    }
     if(is.null(bias_parms))
         bias_parms <- c(1, 1, 1, 1)
     else bias_parms <- bias_parms
@@ -109,6 +116,31 @@ misclassification <- function(case,
         corr.rr <- (A/(A + C)) / (B/(B + D))
         corr.or <- (A/B) / (C/D)
 
+        mle.corr.or <- ((a+((a+b)*(bias_parms[3]-1))) * (((c+d)*bias_parms[2])-c)) /
+            ((c+((c+d)*(bias_parms[4]-1))) * (((a+b)*bias_parms[1])-a))
+
+        se.corr.or <- sqrt(
+        (
+            (
+                (a+b)*a*b*((bias_parms[1]+bias_parms[3]-1)^2)
+            ) /
+            (
+                ((((a+b)*bias_parms[1])-a)^2) * ((((a+b)*bias_parms[3])-b)^2)
+            )
+        ) +
+        (
+            (
+                (c+d)*c*d*((bias_parms[2]+bias_parms[4]-1)^2)
+            ) /
+            (
+                ((((c+d)*bias_parms[2])-c)^2) * ((((c+d)*bias_parms[4])-d)^2)
+            )
+        )
+        )
+        lci.corr.or <- exp(log(mle.corr.or) - qnorm(1 - alpha/2) * se.corr.or)
+        uci.corr.or <- exp(log(mle.corr.or) + qnorm(1 - alpha/2) * se.corr.or)
+        
+
         if (is.null(rownames(tab)))
             rownames(tab) <- paste("Row", 1:2)
         if (is.null(colnames(tab)))
@@ -128,10 +160,12 @@ misclassification <- function(case,
         colnames(rmat) <- c(" ",
                             paste(100 * (alpha/2), "%", sep = ""),
                             paste(100 * (1 - alpha/2), "%", sep = ""))
-        rmatc <- rbind(corr.rr, corr.or)
+        rmatc <- rbind(c(corr.rr, NA, NA), c(corr.or, lci.corr.or, uci.corr.or))
         rownames(rmatc) <- c("Misclassification Bias Corrected Relative Risk:",
                              "   Misclassification Bias Corrected Odds Ratio:")
-        colnames(rmatc) <- " "
+        colnames(rmatc) <- c(" ",
+                            paste(100 * (alpha/2), "%", sep = ""),
+                            paste(100 * (1 - alpha/2), "%", sep = ""))
     }
     
     if (type == "outcome"){
@@ -168,55 +202,6 @@ misclassification <- function(case,
             rownames(corr.tab) <- row.names(tab)
         }
         if (is.null(colnames(tab))){ 
-            colnames(corr.tab) <- paste("Col", 1:2)
-        } else {
-            colnames(corr.tab) <- colnames(tab)
-        }
-        rmat <- rbind(c(obs.rr, lci.obs.rr, uci.obs.rr), c(obs.or, lci.obs.or, uci.obs.or))
-        rownames(rmat) <- c("Observed Relative Risk:", "   Observed Odds Ratio:")
-        colnames(rmat) <- c(" ",
-                            paste(100 * (alpha/2), "%", sep = ""),
-                            paste(100 * (1 - alpha/2), "%", sep = ""))
-        rmatc <- rbind(corr.rr, corr.or)
-        rownames(rmatc) <- c("Misclassification Bias Corrected Relative Risk:",
-                             "   Misclassification Bias Corrected Odds Ratio:")
-        colnames(rmatc) <- " "
-    }
-
-    if (type == "confounder") {
-        obs.rr <- (a/(a + c)) / (b/(b + d))
-        se.log.obs.rr <- sqrt((c/a) / (a+c) + (d/b) / (b+d))
-        lci.obs.rr <- exp(log(obs.rr) - qnorm(1 - alpha/2) * se.log.obs.rr)
-        uci.obs.rr <- exp(log(obs.rr) + qnorm(1 - alpha/2) * se.log.obs.rr)
-
-        obs.or <- (a/b) / (c/d)
-        se.log.obs.or <- sqrt(1/a + 1/b + 1/c + 1/d)
-        lci.obs.or <- exp(log(obs.or) - qnorm(1 - alpha/2) * se.log.obs.or)
-        uci.obs.or <- exp(log(obs.or) + qnorm(1 - alpha/2) * se.log.obs.or)
-
-        Ac1 <- (a - (1 - bias_parms[3]) * (a + b)) / (bias_parms[1] - (1 - bias_parms[3]))
-        C <- (c - (1 - bias_parms[4]) * (c + d)) / (bias_parms[2] - (1 - bias_parms[4]))
-        B <- (a + b) - A
-        D <- (c + d) - C
-
-        if(A < 1 | B < 1 | C < 1 | D < 1)
-            stop('Parameters chosen lead to negative cell(s) in adjusted 2x2 table.')
-        
-        corr.tab <- matrix(c(A, B, C, D), nrow = 2, byrow = TRUE)
-
-        corr.rr <- (A/(A + C)) / (B/(B + D))
-        corr.or <- (A/B) / (C/D)
-
-        if (is.null(rownames(tab)))
-            rownames(tab) <- paste("Row", 1:2)
-        if (is.null(colnames(tab)))
-            colnames(tab) <- paste("Col", 1:2)
-        if (is.null(rownames(tab))){
-            rownames(corr.tab) <- paste("Row", 1:2)
-        } else {
-            rownames(corr.tab) <- row.names(tab)
-        }
-        if (is.null(colnames(tab))){
             colnames(corr.tab) <- paste("Col", 1:2)
         } else {
             colnames(corr.tab) <- colnames(tab)